# khronos: Use vchiq_get_client_id to determine global PID.