@pelwell 2016-06-27
I'm happy with the motivation behind the patch, and I think the implementation is correct. My hesitation has been on the grounds of a potential performance hit, but that doesn't seem to be a problem; if anything it seems marginally quicker:
```
Without:
1260 frames 17.4 seconds 72.6 fps 5.0/13.8/131.0/5.7 ms
1260 frames 17.1 seconds 73.8 fps 5.0/13.6/36.0/4.5 ms
1260 frames 17.2 seconds 73.4 fps 6.0/13.6/37.0/4.5 ms
1260 frames 17.1 seconds 73.7 fps 6.0/13.6/35.0/4.6 ms


With:
1260 frames 17.2 seconds 73.2 fps 5.0/13.7/133.0/5.7 ms
1260 frames 17.1 seconds 73.9 fps 5.0/13.5/36.0/4.5 ms
1260 frames 17.1 seconds 73.8 fps 6.0/13.6/36.0/4.5 ms
1260 frames 17.1 seconds 73.8 fps 5.0/13.5/36.0/4.5 ms
```
I'm happy for this to be merged - @popcornmix ?